bitmap_mgr: rework to a doubly linked list representation
Previously all the tile data was in one large linked list, but that was
just to get a prototype up and running. Move towards a better scheme:
each zoom level has a linked list of "x" rows, and each "x" row has a
linked list of "y" tiles. This allows us to quickly (?) fetch tiles from
the cache. May not be the most optimal data structure, but works.