bitmap_find_cache: start search from last known position
Assuming that bitmap_find_cache() is almost always called in a loop with
incrementing x and y, we can start searching from the last known tile
most of the time. Previously we always walked the whole x and y lists
till we found a match. Now we start searching from the position of the
last tile, if available and reliable. In the ideal case, we're looking
for the next tile, and we take a very fast code path.