bitmap_mgr: cache purge methods: scan lists at current zoomlevel
If we failed to reach our cache target by dropping other zoom levels, we
need to scan the x and y tiles in the current zoom level and delete the
"farthest" ones until we reach the target.

This is not really balanced, since we will delete x rows till just a
narrow band remains, and only then look at the y rows. That's a
consequence of the datastructure design.