xylist: introduce this datastructure
Remove the custom xylist in the bitmap_mgr and replace it with a
generic data structure implementation that we can also use for the
texture manager and other tile cache levels. An xylist is a kind of
linked list that (currently) looks like this:

  zoomlevel ->  x -> x -> x -> x -> x -> x -> x -> x
                |    |    |    |    |    |    |    |
                y    y    y    y    y    y    y    y
                |    |    |         |    |    |
                y    y    y         y    y    y
                |    |              |
                y    y              y

To find a tile at (xn,yn), first traverse the x linked list, and when
you find the element with x->n == xn, traverse its y linked list until
you find the element with y->n == yn. The "tile" is a void pointer to an
opaque data structure (private to caller).

This xylist data structure lets us insert new tiles quickly at a given
coordinate. It also lets us remove "unused" tiles quickly by purging
zoom levels, x rows and y elements that are out of view.

When creating an xylist, the caller supplies a pointer to a callback
function that can procure a data pointer for (zoom, xn, yn). When a tile
is requested, and it cannot be found in the list, the callback is called
to procure the tile data. When the cache cleanup runs, the destructor
callback is called to flush the data. This allows the xylist to be
"generic", it deals with storing pointers at a (zoom,x,y) coordinate. In
a sense it's a kind of database.

We can also do "chaining" of lists. The texture manager procures its
bitmaps from the bitmap manager. The bitmap manager procures its tiles
from the disk manager (not implemented yet). The disk manager procures
its tiles from the online tile servers. Requests "cascade" down from
really fast cache levels (OpenGL texture cache) to really slow levels
(network lookup). The "search depth" parameter determines at which level
the search will terminate.

When we want a set of tiles to cover the screen, we start at the first y
at first x, and end at the last y at the last x. By saving the x and y
pointers used for the last lookup, and using them at the next lookup
(checking that they are appropriate, of course), we can retrieve
elements really fast, almost as fast as a sequential lookup.