xylist: scan forward or reverse for reasonable start position
Previously, if we did not have a good saved (x,y) coordinate, we would
automatically start scanning from the start of the list, that is to say,
xs or ys. Now we compare the distance from the start node with the
distance from the end node, and choose the "closest" one to start our
search from. If it's the end node, we run the linked list backwards.