xylist_insert_tile: add public function for async tile fetchers
Lets the list owner insert a tile into the list at any time.
This can be used by nonblocking tile fetchers. The owner asks the xylist
for a tile, the xylist calls the procure function, the procure function
immediately returns NULL and shoots off a thread that gets the tile
asynchronously in the background. That thread then uses this function
to add the tile to the list when it's done.