bitmap_mgr: add threaded disk lookups
After much sweat and tears, we have a system that does threaded disk
lookups, meaning that when we want a tile from disk, we fire off a
thread to go get it for us in the background. This keeps the user
interface fast. Eventually we will implement the same thing for network
lookups.

Threads are administrated from their own xylist, which means they
are amenable to auto-creation and auto-deletion just as tiles are. To
facilitate this, we need to be able to cancel pthreads, which turned out
to be more difficult than it looks if you don't want to leak file
descriptors and other resources. In the end we have something that meets
the requirements, but is still a bit messy. To be cleaned up later.