pngloader: use separate control structure to check on thread
Threads don't have to delete themselves from the xylist, or hang around
occupying the slot for a few seconds. Each thread has a control node
with its run state and a pointer to its thread ID, a pointer to which is
saved for as long as we can manage in the threadlist. If the tile has a
node, we know we've tried it previously.