Use OpenGL pixel functions directly: massive speedup
Do not draw textured quads; textures are really slow if you have to
create and destroy them for each tile, and video memory is too small for
adequate caching. Instead, use the pixel functions to draw bitmaps
straight from our large RAM buffer. This results in a massively faster
interface, and alloes us to get rid of the finicky texture_mgr module,
and with that, a layer of indirection.

Also remove the tile shader, since we render our pixels directly now.