layer_blanktile: factor blank tile code into separate layer
Do not draw blank tiles as a part of the OSM layer, but draw them in
their own separate layer.

This is marginally less efficient since we now always draw the blank
tiles, even when they will be occluded by an overlaying OSM tile. Leave
these kinds of optimization for later.