Background, blank tiles: use OpenGL primitives, not textures
Draw the background as a set of OpenGL lines in a fixed viewport instead
of using a slow fragment shader to get the exact same effect.

Draw the blank tiles and the lines between them procedurally with OpenGL
quads and lines, instead of blitting the same thing to the screen as a
texture.

Split the OpenGL setup code in the viewport into a part that sets up the
viewport for 1:1 drawing to screen coordinates, and a part that sets the
viewport to world (tile) coordinates.