Viewport: calculate frustum and its bounding box in world
Get the frustum by reverse-projecting the screen corners, and ensure
that we deal with an infinite (horizon) view properly. Right now we
solve it by calculating the far ends of the frustum ourselves and
chopping them off at an arbitrary distance.

Also calculate the frustum's bounding box inside the world. This will
form the basis for all sorts of tile calculations. It also turns out to
be far from trivial to reliably intersect a polygon with a square.