Viewport: precalculate frustum barycentric coordinates
Precalculate all the values we can precalculate, so that the check of
whether a point is inside the frustum can be a bit more efficient.

Also, when deciding the bounding box, use this frustum check whenever
one of the frustum ribs is outside the world. More efficient and correct
than the previous method of forming two overlapping triangles from the
out-of-bounds rib.