Tilepicker: introduce separate tile-picking engine
Previously, layer_osm contained some of the logic of how to pick a tile
for a certain tile location. Move that responsibility to a separate
routine; layer_osm just gets the draw list when it asks for it.

Layer_overview renders all tiles and their zoom level for debug
purposes.