Tiledrawer: add tile subdivision to ensure globe roundness
When a tile is larger than a certain fraction of the map width (1/16
currently), subdivide it into multiple sub-tiles of at most 1/16 the map
width. This guarantees the "roundness" of the globe at low resolutions:
at zoom level 0, the "globe" would be a point, at zoom level 1 a flat
diamond, and at zoom level 2, a crude diamond.