tilepicker.c: use floats for tile coordinates
Use floats for coordinates so we can postulate "fractional" tiles.
This will be useful later when we want to draw the globe with a
guaranteed number of subdivisions at very low zoom levels, to keep
the spherical character when the world is just one tile across.

Floats should be fine because they can represent multiples of two
without loss of precision, and all we do is halving, doubling, and
dividing by powers of two. We use ldexpf() for losslessly
manipulating the exponent.