Autoscroll: overhaul, use center coordinates
Autoscroll is based on the coordinates of the center marker. Overhaul
the code to make that more obvious. The old code based the autoscroll
amount on the pixel coordinates of the mouse, which is inaccurate and
leads to a lot of projecting.

Also change the movement code to use a position plus a speed.