tilepicker: use viewing angle in zoom calculation
The level of detail should change if the viewing angle or the window
width changes. One window pixel should always show the same level of
detail, taking into account camera distance, viewing angle and viewport
width.

Achieve this by calculating the perpendicular "ground distance" swept by
each pixel, and using that as a metric for the world level of detail (in
other words, tile zoom level) that the pixel ought to display.