cache: support cache entries of variable size
Support cache entries with variable size. This means that the cache can
contain many different items, from single pointers to very large
structures. This obviates the need for users to construct their own
schemes on top of the cache to store large or complex entries. It also
keeps the size of the "core" cache_node structure down.

Internally, the entries are stored inside a linear array which is
indexed by the cache nodes.

This commit does not compile on its own.