layer.h: split layers.h into per-layer and all-layer parts
Split the `layers.h' header file into a file that contains only
definitions that apply to one layer (`layer.h'), and a file that
contains only definitions that apply to the list of all layers.

This is a cleanup commit that will make it easier to make some necessary
changes in the layer code.