program.h: split programs.h into per-program and all-program parts
Split the `programs.h' header file into a file that contains only
definitions that apply to one program (`program.h'), and a file that
contains only definitions that apply to the list of all programs.

This is a cleanup commit that will make it easier to make some necessary
changes in the program code.